- Fixed bug #014635: For a discount with decimal component shop/basket cuts off the decimal (via Gunnstein), merged from trunk rev 23394
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/4.0@23396 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/classes/ezproductcollectionitem.php
@@ -84,7 +84,7 @@ class eZProductCollectionItem extends eZPersistentObject
                                                                'default' => 0,
                                                                'required' => true ),
                                          'discount' => array( 'name' => "DiscountValue",
-                                                              'datatype' => 'integer',
+                                                              'datatype' => 'float',
                                                               'default' => 0,
                                                               'required' => true ),
                                          'name' => array( 'name' => 'Name',
